Rename packages and introduce interfaces
