--- src/main/java/com/hedvig/client/dto/Key.kt
@@ -1,5 +1,0 @@
-package com.hedvig.client.dto
-
-data class Key(
-    val value: String
-)
