+++ src/main/java/com/hedvig/localization/client/dto/Key.kt
@@ -1,0 +1,5 @@
+package com.hedvig.localization.client.dto
+
+data class Key(
+    val value: String
+)
