+++ src/main/java/com/hedvig/localization/client/dto/LocalizationData.kt
@@ -1,4 +1,4 @@
-package com.hedvig.client.dto
+package com.hedvig.localization.client.dto
 
 data class LocalizationData(
     val languages: List<Language>
