+++ src/main/java/com/hedvig/localization/client/dto/Translation.kt
@@ -1,4 +1,4 @@
-package com.hedvig.client.dto
+package com.hedvig.localization.client.dto
 
 data class Translation(
     val key: Key?,
