+++ src/main/java/com/hedvig/localization/service/LocalizationService.kt
@@ -1,0 +1,7 @@
+package com.hedvig.localization.service
+
+import java.util.Locale
+
+interface LocalizationService {
+  fun getText(locale: Locale?, key: String): String?
+}
