+++ src/main/java/com/hedvig/localization/service/TextKeysLocaleResolver.kt
@@ -1,0 +1,7 @@
+package com.hedvig.localization.service
+
+import java.util.Locale
+
+interface TextKeysLocaleResolver {
+    fun resolveLocale(acceptLanguage: String?): Locale
+}
